feat: Criando classe Profissional e linkando ao Usuário

+++ src/main/java/com/ian/projetointegradoianbs/repository/ProfissionalRepository.java
@@ -7,5 +7,5 @@ import com.ian.projetointegradoianbs.domain.Profissional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfissionalRepository extends JpaRepository<Profissional, Long> {
-    public Optional<Profissional> findByUsuario(Long usuarioId);
+    public Optional<Profissional> findByUsuario(Integer usuarioId);
 }
